refactor klasy Stack
