+++ src/main/java/edu/kis/vh/stacks/StackHanoi.java
@@ -1,6 +1,6 @@
 package edu.kis.vh.stacks;
 
-public class StackHanoi extends stack {
+public class StackHanoi extends Stack {
 
 int totalRejected = 0;
