+++ src/test/java/edu/kis/vh/stacks/unittests/stackTest.java
@@ -3,13 +3,13 @@ package edu.kis.vh.stacks.unittests;
 import org.junit.Assert;
 import org.junit.Test;
 
-import edu.kis.vh.stacks.stack;
+import edu.kis.vh.stacks.Stack;
 
 public class stackTest {
 	
 	@Test
 	public void testPush() {
-		stack stackObj = new stack();
+		Stack stackObj = new Stack();
 		int testValue = 4;
 		stackObj.push(testValue);
 		
@@ -19,7 +19,7 @@ public class stackTest {
 
 	@Test
 	public void testIsEmpty() {
-		stack stackObj = new stack();
+		Stack stackObj = new Stack();
 		boolean result = stackObj.isEmpty();		
 		Assert.assertEquals(true, result);
 		
@@ -31,7 +31,7 @@ public class stackTest {
 
 	@Test
 	public void testIsFull() {
-		stack stackObj = new stack();
+		Stack stackObj = new Stack();
 		final int STACK_CAPACITY = 12;
 		for (int i = 0; i < STACK_CAPACITY; i++) {
 			boolean result = stackObj.isFull();		
@@ -45,7 +45,7 @@ public class stackTest {
 
 	@Test
 	public void testTop() {
-		stack stackObj = new stack();
+		Stack stackObj = new Stack();
 		final int EMPTY_STACK_VALUE = -1;
 		
 		int result = stackObj.top();
@@ -62,7 +62,7 @@ public class stackTest {
 
 	@Test
 	public void testPop() {
-		stack stackObj = new stack();
+		Stack stackObj = new Stack();
 		final int EMPTY_STACK_VALUE = -1;
 		
 		int result = stackObj.pop();
